+++ examples/array_contains.php
@@ -1,5 +1,5 @@
 <?php
-prggmr\load_signal('arrays');
+prggmr\load_module('arrays');
 
 prggmr\handle(new \prggmr\signal\arrays\Contains([0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10]), function($value){
     echo "This used a array search";