+++ examples/cron.php
@@ -1,19 +1,8 @@
 <?php
 date_default_timezone_set('UTC');
-prggmr\load_signal('time');
 
-
+prggmr\load_module('cron');
 // Run every minute
-prggmr\signal\time\cron('* * * * *', function(){
+prggmr\module\cron\cron('* * * * *', function(){
     echo "EVERY MINUTE!".PHP_EOL;
 });
-
-prggmr\handle('signal', function(){
-
-});
-
-$handle = new prggmr\Handle(function(){
-
-}, 1, 0);
-
-prggmr\handle('signal', $handle);