+++ examples/env.php
@@ -1,6 +1,0 @@
-#!/usr/bin/env prggmr
-<?php
-prggmr\interval(function(){
-    $a = new stdClass();
-    echo "1 second";
-}, 1000);