+++ examples/loader.php
@@ -1,5 +1,5 @@
 <?php
-prggmr\load_signal('string');
+prggmr\load_module('string');
 
 prggmr\handle_loader('test_hey', dirname(realpath(__FILE__))."/loader");