+++ examples/performance.php
@@ -2,9 +2,8 @@
 ini_set('memory_limit', -1);
 $time = microtime(true);
 echo "Start".PHP_EOL;;
-for ($i=0;$i!=1000;$i++){
+for ($i=0;$i!=100;$i++){
     prggmr\handle($i, function(){
-        $this->a++;
     });
 }
 echo "Handle Register".PHP_EOL;
@@ -13,7 +12,7 @@ echo PHP_EOL;
 $event = new \prggmr\Event();
 $event->a = 1;
 $time = microtime(true);
-for ($i=0;$i!=1000;$i++){
+for ($i=0;$i!=100;$i++){
     prggmr\signal($i, $event);
 }
 echo $event->a.PHP_EOL;